Settings: request runtime permissions inside VoiceWakeup
If the user has selected a Direct dial shortcut for their voice wakeup
action, then the app which launches this intent must hold the dangerous
permission CALL_PHONE.

So when the user picks an intent which would require this permission,
make sure VoiceWakeup requests and grants the permission. Otherwise
reset to default.

Ref: CYNGNOS-1560

Change-Id: I02e2c7a7d9958a5b15f438687f509802e75db704
Signed-off-by: Roman Birg <roman@cyngn.com>